Accepted

+++ Sqrt(x).cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        float r=sqrt(x);
-        int ans= floor(r);
+        double r=sqrt(x);
+        double ans= floor(r);
         return ans;
     }
 };
